Split share extension in multiple files
