+++ Wire-iOS Share Extension/ShareViewController.swift
@@ -325,173 +325,3 @@ extension ShareViewController {
         }
     }
 }
-
-// MARK: - Sending progress
-class SendingProgressViewController : UIViewController, SendableObserver {
-    
-    var sentHandler : (() -> Void)?
-    var cancelHandler : (() -> Void)?
-    
-    private var progressLabel = UILabel()
-    private var observers : [(Sendable, SendableObserverToken)] = []
-    
-    var totalProgress : Float {
-        var totalProgress : Float = 0.0
-        
-        observers.forEach { (message, _) in
-            if message.deliveryState == .sent || message.deliveryState == .delivered {
-                totalProgress = totalProgress + 1.0 / Float(observers.count)
-            } else {
-                let messageProgress = (message.deliveryProgress ?? 0)
-                totalProgress = totalProgress +  messageProgress / Float(observers.count)
-            }
-        }
-        
-        return totalProgress
-    }
-    
-    var isAllMessagesDelivered : Bool {
-        return observers.reduce(true) { (result, observer) -> Bool in
-            return result && (observer.0.deliveryState == .sent || observer.0.deliveryState == .delivered)
-        }
-    }
-    
-    init(messages: [Sendable]) {
-        super.init(nibName: nil, bundle: nil)
-        
-        messages.forEach {message in
-            observers.append((message, (message.registerObserverToken(self))))
-        }
-    }
-    
-    deinit {
-        observers.forEach { (message, token) in
-            message.remove(token)
-        }
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        self.navigationItem.hidesBackButton = true
-        self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Cancel", style: .plain, target: self, action: #selector(onCancelTapped))
-        
-        progressLabel.text = "0%";
-        progressLabel.textAlignment = .center
-        progressLabel.font = UIFont.systemFont(ofSize: 32)
-        
-        view.addSubview(progressLabel)
-        
-        constrain(view, progressLabel) { container, progressLabel in
-            progressLabel.edges == container.edgesWithinMargins
-        }
-    }
-    
-    func onCancelTapped() {
-        observers.filter {
-            $0.0.deliveryState != .sent && $0.0.deliveryState != .delivered
-        }.forEach {
-            $0.0.cancel()
-        }
-        cancelHandler?()
-    }
-    
-    func onDeliveryChanged() {
-        progressLabel.text = "\(Int(self.totalProgress * 100))%"
-        
-        if self.isAllMessagesDelivered {
-            sentHandler?()
-        }
-    }
-}
-
-class NotSignedInViewController : UIViewController {
-    
-    var closeHandler : (() -> Void)?
-    
-    let messageLabel = UILabel()
-    
-    init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        self.navigationItem.hidesBackButton = true
-        self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Close", style: .plain, target: self, action: #selector(onCloseTapped))
-        
-        messageLabel.text = "You need to sign into Wire before you can share anything";
-        messageLabel.textAlignment = .center
-        messageLabel.numberOfLines = 0
-        
-        view.addSubview(messageLabel)
-        
-        constrain(view, messageLabel) { container, messageLabel in
-            messageLabel.edges == container.edgesWithinMargins
-        }
-    }
-    
-    func onCloseTapped() {
-        closeHandler?()
-    }
-}
-
-class ConversationSelectionViewController : UITableViewController {
-    
-    var conversations : [Conversation]
-    
-    var selectionHandler : ((_ conversation: Conversation) -> Void)?
-    
-    init(conversations: [Conversation]) {
-        self.conversations = conversations
-        
-        super.init(style: .plain)
-        
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "ConversationCell")
-        preferredContentSize = UIScreen.main.bounds.size
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .clear
-    }
-    
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return conversations.count
-    }
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let conversation = conversations[indexPath.row]
-        let cell = tableView.dequeueReusableCell(withIdentifier: "ConversationCell", for: indexPath)
-        
-        cell.textLabel?.text = conversation.name
-        cell.backgroundColor = .clear
-        
-        return cell
-    }
-    
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        if let selectionHandler =  selectionHandler {
-            selectionHandler(conversations[indexPath.row])
-        }
-    }
-    
-}
